refactors navbar
